# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SaveUpdateResProduct": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Price": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"Name",
					"Price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqProduct": {
				"properties": {
					"Name": {
						"type": "string"
					},
					"Price": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"Name",
					"Price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetProduct": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqProduct": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Price": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"Name",
					"Price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteProduct": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "productslatestp",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/product/saveproduct": {
			"post": {
				"operationId": "Saveproduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResProduct"
								}
							}
						}
					}
				},
				"tags": [
					"product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqProduct"
							}
						}
					}
				}
			}
		},
		"/product/getproduct": {
			"post": {
				"operationId": "Getproduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResProduct"
								}
							}
						}
					}
				},
				"tags": [
					"product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetProduct"
							}
						}
					}
				}
			}
		},
		"/product/updateproduct": {
			"put": {
				"operationId": "Updateproduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResProduct"
								}
							}
						}
					}
				},
				"tags": [
					"product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqProduct"
							}
						}
					}
				}
			}
		},
		"/product/deleteproduct": {
			"delete": {
				"operationId": "Deleteproduct",
				"responses": {
					"200": {
						"description": "Product Deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResProduct"
								}
							}
						}
					}
				},
				"tags": [
					"product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteProduct"
							}
						}
					}
				}
			}
		},
		"/product/getproductlist": {
			"post": {
				"operationId": "Getproductlist",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResProduct"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"product"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}